// src/apis/paises/interfaces/paises.response.interface.ts
export interface PaisesResponse {
  id: string;
  nombre: string;
  code: string;
  code_phone: string;
  nombre_moneda: string;
  simbolo_moneda: string;
  nombre_documento: string;
  isActive: boolean;
  departamentos: Departamento[];
}

export interface Departamento {
  id: string;
  nombre: string;
  municipios: Municipio[];
}

export interface Municipio {
  id: string;
  nombre: string;
}